Update endpoints, GET:drinks-detail, DELETE:drinks, POST:drinks

+++ backend/src/api.py
@@ -37,43 +37,49 @@ db_drop_and_create_all()
     returns status code 200 and json {"success": True, "drinks": drinks} where drinks is the list of drinks
         or appropriate status code indicating reason for failure
 '''
-@app.route('/drinks')
-def get_drinks():
-    drinks =[]
-    available_drinks = Drink.query.order_by(Drink.id).all()
-    for drink in available_drinks:
-        drinks.append(drink.short())
-    if len(drinks) == 0:
-        abort(404)
+# Public endpoint for getting drinks
 
-    return jsonify({'success': True,
-                    'drinks': drinks
-                    }), 200
+@app.route('/drinks',methods=['GET'])
+@requires_auth('get:drinks')
+def get_drinks(payload):
+    try:
+        available_drinks = Drink.query.all()
+        drinks = [drink.short() for drink in available_drinks]        
+        if payload:
+            return jsonify({                
+                'success':True,
+                'drinks':drinks
+            })
+        else:
+            abort(403)
+    except AuthError:
+        abort(401)
 
 '''
-@TODO implement endpoint
+@TODO implement endpoint (DONE)
     GET /drinks-detail
         it should require the 'get:drinks-detail' permission
         it should contain the drink.long() data representation
     returns status code 200 and json {"success": True, "drinks": drinks} where drinks is the list of drinks
         or appropriate status code indicating reason for failure
 '''
+
 @app.route('/drinks-detail')
 @requires_auth('get:drinks-detail')
 def get_drink_detail(payload):
-    print(payload)
-    drinks = []
-    available_drinks = Drink.query.order_by(Drink.id).all()
-    for drink in available_drinks:
-        drinks.append(drink.long())
-    if len(drinks) == 0:
-        abort(404)
-    return jsonify({'success': True,
-                    'drinks': drinks,
+    try:
+        available_drinks = Drink.query.all()
+        drinks = [drink.long() for drink in available_drinks]
+
+        return jsonify({'success': True,
+                    'drinks': drinks
                     }), 200
+    except Exception as e:
+        print(e)
+        abort(404)
 
 '''
-@TODO implement endpoint
+@TODO implement endpoint (DONE)
     POST /drinks
         it should create a new row in the drinks table
         it should require the 'post:drinks' permission
@@ -84,9 +90,9 @@ def get_drink_detail(payload):
 '''
 @app.route('/drinks', methods=['POST'])
 @requires_auth('post:drinks')
-def add_drink(jwt):
+def add_drink(payload):
     body = request.get_json()
-    print(body)
+    # print(body)
 
 # New Data
     new_title = body.get('title', None)
@@ -102,13 +108,12 @@ def add_drink(jwt):
         return jsonify({'success': True,
                         'drinks': [drink.long()],
                         }), 200
-    except BaseException:
-        # print(sys.exc_info())
+    except:
+        print(sys.exc_info())
         abort(422)
 
-
 '''
-@TODO implement endpoint
+@TODO implement endpoint (DONE)
     PATCH /drinks/<id>
         where <id> is the existing model id
         it should respond with a 404 error if <id> is not found
@@ -121,12 +126,12 @@ def add_drink(jwt):
 
 @app.route('/drinks/<int:drink_id>', methods=['PATCH'])
 @requires_auth('patch:drinks')
-def update_drink(payload, id):
+def update_drink(payload, drink_id):
     
     body = request.get_json()
     try:
 
-        drink = Drink.query.filter(Drink.id == id).one_or_none()
+        drink = Drink.query.filter(Drink.id == drink_id).one_or_none()
         if drink is None:
             abort(404)
         
@@ -146,10 +151,8 @@ def update_drink(payload, id):
         print(sys.exc_info())
         abort(422)
 
-
-
 '''
-@TODO implement endpoint
+@TODO implement endpoint (DONE)
     DELETE /drinks/<id>
         where <id> is the existing model id
         it should respond with a 404 error if <id> is not found
@@ -158,26 +161,31 @@ def update_drink(payload, id):
     returns status code 200 and json {"success": True, "delete": id} where id is the id of the deleted record
         or appropriate status code indicating reason for failure
 '''
+
 @app.route('/drinks/<int:drink_id>', methods=['DELETE'])
 @requires_auth('delete:drinks')
-def delete_drink(payload, drink_id):
-    drink = Drink.query.filter(Drink.id == drink_id).one_or_none()
-    if drink is None:
-        abort(404)
+def delete_drink(payload,drink_id):
     try:
-        drink.delete()
-        return jsonify({'success': True,
-                        'delete': drink_id,
-                        }), 200
-    except BaseException:
-        abort(422)
-
+        drink = Drink.query.filter(Drink.id == drink_id).one_or_none()
+        if drink is None:
+            abort(404)
+        else:
+            if payload:
+                drink.delete()
+                return jsonify({
+                    'success':True,
+                    'delete': drink_id
+                })
+            else:
+                abort(403)
+    except AuthError:
+        print(sys.exc_info())
+        abort(401)
 
 # Error Handling
 '''
 Example error handling for unprocessable entity
 '''
-
 @app.errorhandler(422)
 def unprocessable(error):
     return jsonify({
@@ -185,7 +193,6 @@ def unprocessable(error):
         "error": 422,
         "message": "unprocessable"
     }), 422
-
 
 '''
 @TODO implement error handlers using the @app.errorhandler(error) decorator
@@ -199,7 +206,7 @@ def unprocessable(error):
 '''
 
 '''
-@TODO implement error handler for 404
+@TODO implement error handler for 404 (DONE)
     error handler should conform to general task above
 '''
 @app.errorhandler(404)
